룸리스트 코드를 뷰컨트롤러에 옮김
